+++ src/index.ts
@@ -5,3 +5,4 @@ export { default as FhiTable } from './components/FhiTable';
 export { default as FhiHeader } from './components/FhiHeader';
 export { default as FhiTabBar } from './components/FhiTabBar';
 export { default as FhiAccordion } from './components/FhiAccordion';
+export { default as FhiToastProvider, useToast as useFhiToast } from './components/FhiToast';